backend/requirements.txt
# MindBridge Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning & AI
torch>=2.6.0
torchvision>=0.17.0
transformers>=4.36.2
datasets==2.15.0
accelerate==0.25.0
opacus==1.4.0

# Privacy & Cryptography
cryptography>=45.0.0
pynacl>=1.5.0
python-gnupg>=0.5.2

# Data Processing
numpy>=1.24.4
pandas>=2.1.4
scipy>=1.11.4
scikit-learn>=1.3.2

# Database & Storage
sqlalchemy>=2.0.23
sqlite3-fts4>=0.0.6
aiosqlite>=0.19.0

# HTTP & Networking
httpx>=0.25.2
aiofiles>=23.2.1
python-multipart>=0.0.6

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Logging & Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
typer>=0.9.0
rich>=13.7.0
tqdm>=4.66.1

# Date & Time
python-dateutil>=2.8.2
pytz>=2023.3

# Text Processing
nltk>=3.8.1
spacy>=3.7.2
textblob>=0.17.1

# Model Optimization
onnx>=1.15.0
onnxruntime>=1.16.3

# Cultural Adaptation
polyglot>=16.7.4
langdetect>=1.0.9

# Health & Wellness APIs
requests>=2.31.0
beautifulsoup4>=4.12.2